monetary values are now using BigDecimal

+++ src/main/java/CarRental/FakeData.java
@@ -5,6 +5,8 @@ import CarRental.domain.Client;
 import CarRental.repositories.CarRepository;
 import CarRental.repositories.UserRepository;
 
+import java.math.BigDecimal;
+
 public class FakeData {
 
     public static void addClients(){
@@ -22,11 +24,11 @@ public class FakeData {
     }
 
     public static void addCars(){
-        Car ferrari = new Car("Ferrari", "F-50", "abc1234", 5000);
+        Car ferrari = new Car("Ferrari", "F-50", "abc1234", BigDecimal.valueOf(5000));
         CarRepository.addCar(ferrari);
-        Car palio = new Car("Fiat", "Palio", "abc2345", 100);
+        Car palio = new Car("Fiat", "Palio", "abc2345", BigDecimal.valueOf(100));
         CarRepository.addCar(palio);
-        Car pegeout = new Car("Pegeout", "206", "abc3456", 150);
+        Car pegeout = new Car("Pegeout", "206", "abc3456", BigDecimal.valueOf(150));
         CarRepository.addCar(pegeout);
 
     }
